Move AppLogo to atoms
